- Initial crack at generalizing focus in LayoutBase - Renamed Create -> Show, Destroy -> Hide - Added IDisplayable interface for Layouts and Windows - A couple other small fixes.

+++ src/musikbox/CommandWindow.cpp
@@ -55,7 +55,7 @@ void CommandWindow::Focus() {
     wmove(this->GetContent(), 0, bufferPosition);
 }
 
-void CommandWindow::WriteChar(int ch) {
+void CommandWindow::WriteChar(int64 ch) {
     if (bufferPosition >= MAX_SIZE) {
         return;
     }
