+++ src/musikbox/IInput.h
@@ -4,6 +4,6 @@
 
 class IInput {
     public:
-        virtual void WriteChar(int ch) = 0;
+        virtual void WriteChar(int64 ch) = 0;
         virtual void Focus() = 0;
 };
